+++ CHANGELOG.md
@@ -18,8 +18,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Warn about evaluation being different from Vidore, and do not store results to prevent confusion.
 - Remove duplicate resize code in `ColQwen2Processor` and `ColQwen2_5_Processor`.
 - Simplify sequence padding for pixel values in `ColQwen2Processor` and `ColQwen2_5_Processor`.
-- Remove deprecated eval from trainer
+- Remove deprecated evaluation (`CustomRetrievalEvaluator`) from trainer
 - Refactor the collator classes
+- Make `processor` input compulsory in `ColModelTrainingConfig`
+- Make `BaseVisualRetrieverProcessor` inherit from `ProcessorMixin`
+- Remove unused `tokenizer` field from `ColModelTrainingConfig`
 
 ## [0.3.8] - 2025-01-29