+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Allow user to pass custom contextual context for passage inference
+- Allow user to pass custom textual context for passage inference
 - Add ColQwen2.5 support
 - Add support for token pooling with `HierarchicalTokenPooler`.
 - Allow user to specify the maximum number of image tokens in the resized images in `ColQwen2Processor` and `ColQwen2_5_Processor`.